MEBS Figure/requirements.txt
numpy>=1.19.2
opencv-python>=4.5.1
PyQt5>=5.15.4
pandas>=1.2.4
scipy>=1.6.2
matplotlib>=3.4.2
statsmodels>=0.12.2
Pillow>=8.2.0 